+++ crates/vm/wasm/src/memory.rs
@@ -70,108 +70,3 @@ fn write_region(memory: &MemoryView, offset: u32, region: Region) -> VmResult<()
     let wptr = <WasmPtr<Region>>::new(offset);
     wptr.deref(memory).write(region).map_err(Into::into)
 }
-
-// ----------------------------------- tests -----------------------------------
-
-#[cfg(test)]
-mod tests {
-
-    use {super::*, std::mem, test_case::test_case};
-
-    #[test]
-    fn region_has_known_size() {
-        // 3x4 bytes with no padding
-        assert_eq!(mem::size_of::<Region>(), 12);
-    }
-
-    #[test_case(
-        UncheckedRegion {
-            offset: 23,
-            capacity: 500,
-            length: 0,
-        }; "empty"
-    )]
-    #[test_case(
-        UncheckedRegion {
-            offset: 23,
-            capacity: 500,
-            length: 250,
-        }; "half full"
-    )]
-    #[test_case(
-        UncheckedRegion {
-            offset: 23,
-            capacity: 500,
-            length: 500,
-        }; "full"
-    )]
-    #[test_case(
-        UncheckedRegion {
-            offset: u32::MAX,
-            capacity: 0,
-            length: 0,
-        }; "end of linear memory 1"
-    )]
-    #[test_case(
-        UncheckedRegion {
-            offset: 1,
-            capacity: u32::MAX - 1,
-            length: 0,
-        }; "end of linear memory 2"
-    )]
-    fn valid_regions(region: UncheckedRegion) {
-        Region::try_from(region).unwrap();
-    }
-
-    #[test_case(
-        UncheckedRegion {
-            offset: 0,
-            capacity: 500,
-            length: 250,
-        },
-        |error| {
-            assert!(matches!(error, VmError::RegionZeroOffset));
-        };
-        "zero offset"
-    )]
-    #[test_case(
-        UncheckedRegion {
-            offset: 23,
-            capacity: 500,
-            length: 501,
-        },
-        |error| {
-            assert!(matches!(error, VmError::RegionLengthExceedsCapacity { length, capacity } if length == 501 && capacity == 500));
-        };
-        "length exceeding capacity"
-    )]
-    #[test_case(
-        UncheckedRegion {
-            offset: 23,
-            capacity: u32::MAX,
-            length: 501,
-        },
-        |error| {
-            assert!(matches!(error, VmError::RegionOutOfRange { offset, capacity } if offset == 23 && capacity == u32::MAX));
-        };
-        "exceeding address space 1"
-    )]
-    #[test_case(
-        UncheckedRegion {
-            offset: u32::MAX,
-            capacity: 1,
-            length: 0,
-        },
-        |error| {
-            assert!(matches!(error, VmError::RegionOutOfRange { offset, capacity } if offset == u32::MAX && capacity == 1));
-        };
-        "exceeding address space 2"
-    )]
-    fn unvalid_regions<F>(region: UncheckedRegion, callback: F)
-    where
-        F: Fn(VmError),
-    {
-        let error = Region::try_from(region).unwrap_err();
-        callback(error);
-    }
-}
